backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
firebase-admin>=6.5.0
pydantic>=2.7.4
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
email-validator==2.1.1
httpx==0.27.0
pypdf>=4.0.0
# PDF text extraction — 3-library cascade for maximum compatibility:
# pymupdf (primary) → pdfplumber (secondary) → pypdf (fallback)
pymupdf>=1.24.0        # handles Canva/Figma/Word exports, unusual encodings
pdfplumber>=0.11.0     # good for tables & multi-column layouts
# LangChain + AI Providers
langchain-core>=0.3.0
langchain-mistralai>=0.2.0
langchain-openai>=0.2.0     # OpenAI integration (primary provider)
langchain-community>=0.3.0  # For PyPDFLoader and other utilities
openai>=1.0.0               # OpenAI Python SDK

# LangGraph + LangSmith for Multi-Agent System
langgraph>=0.2.0            # Agent orchestration and workflow management
langsmith>=0.1.0            # Tracing, monitoring, and evaluation
pytest==8.2.2
pytest-asyncio==0.23.7
slowapi==0.1.9
cloudinary==1.41.0
PyJWT>=2.8.0
cryptography>=42.0.0
# Error tracking (optional — loaded conditionally when SENTRY_DSN is set)
sentry-sdk>=2.0.0
# OCR support for scanned PDFs (optional — fallback if primary PDF extractors fail)
Pillow>=10.0.0
pytesseract>=0.3.10
